Ajaxify design date page

+++ app/assets/javascripts/date_design.js
@@ -2,8 +2,9 @@ $(document).on('page:change', function() {
 
 	var dateId = $('#selected-experience').attr('data-playdate-id');
 	var experienceId = $('#selected-experience .select-experience-btn').attr('data-experience-id');
-
-	$('#selected-experience .experience-details').prepend('<a href="/dates/' + dateId + '/delete" data-experience-id="'+ experienceId +'" class="selected-experience-delete-btn"><i class="close-icon fa fa-times"></i></a>')
+	if (!$('#recipient-confirmed').hasClass('fa-check-circle') || !$('#initiator-confirmed').hasClass('fa-check-circle')) {
+		$('#selected-experience .experience-details').prepend('<a href="/dates/' + dateId + '/delete" data-experience-id="'+ experienceId +'" class="selected-experience-delete-btn"><i class="close-icon fa fa-times"></i></a>')
+	}
 
 	$('.playdate-container').on('click', '.select-experience-btn a', function(e) {
 		e.preventDefault();
@@ -34,6 +35,13 @@ $(document).on('page:change', function() {
 			};
 			$('#recipient-confirmed').removeClass('fa-check-circle').addClass('fa-circle-thin');
 			$('#initiator-confirmed').removeClass('fa-check-circle').addClass('fa-circle-thin');
+			if ($('.disabled-button')) {
+				if ($('#recipient-accepted').hasClass('fa-check-circle')) {
+					$('.accept-confirm-pay-wrapper').html('<a id="confirm-invitation" class="accept-confirm-pay-btn" href="/dates/' + data.playdate.id + '/confirm">Confirm Date Selection</a>');
+				} else {
+					$('.accept-confirm-pay-wrapper').html('<a id="accept-invitation" class="accept-confirm-pay-btn" href="/dates/' + data.playdate.id + '/accept">Accept Invitation</a>');
+				}
+			};
 		});
 
 	});
@@ -58,11 +66,13 @@ $(document).on('page:change', function() {
 			} else {
 				$('.accept-confirm-pay-wrapper').hide();
 			};
+			$('#recipient-confirmed').removeClass('fa-check-circle').addClass('fa-circle-thin');
+			$('#initiator-confirmed').removeClass('fa-check-circle').addClass('fa-circle-thin');
 		});
 
 	});
 
-	$('#accept-invitation').on('click', function(e) {
+	$('.playdate-container').on('click', '#accept-invitation', function(e) {
 		e.preventDefault();
 
 		$.ajax({
@@ -76,7 +86,7 @@ $(document).on('page:change', function() {
 		});
 	});
 
-	$('#confirm-invitation').on('click', function(e) {
+	$('.playdate-container').on('click', '#confirm-invitation', function(e) {
 		e.preventDefault();
 
 		$.ajax({
@@ -89,7 +99,15 @@ $(document).on('page:change', function() {
 			} else {
 				$('#initiator-confirmed').removeClass('fa fa-circle-thin').addClass('fa fa-check-circle');
 			};
-			$('.accept-confirm-pay-wrapper').html('<div class="disabled-button desaturate italics">Awaiting confirmation from other parent</div>');
+			if ($('#recipient-confirmed').hasClass('fa-check-circle') && $('#initiator-confirmed').hasClass('fa-check-circle')) {
+				$('#confirm-invitation').remove();
+				$('.pay-btn-wrapper').removeClass('hidden');
+				$('.food-and-drink-container').slideUp();
+				$('.movie-container').slideUp();
+				$('.selected-experience-delete-btn').remove();
+			} else {
+				$('.accept-confirm-pay-wrapper').html('<div class="disabled-button desaturate italics">Awaiting confirmation from other parent</div>');
+			}
 		});
 	});
